move tertiary nav off component into styles
