+++ src/docs/foundations/Foundations.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useContext } from 'react';
 import { NavItemSecondary } from '../../library/components/navigation/navSecondary/NavSecondary';
-import { NavItemTertiary } from '../../library/components/navigation/navTertiary/NavTertiary';
+import { NavItemTertiary } from '../../library/components/navigation/NavTertiary';
 import { IdLookupContext } from '../App';
 import Markdown from '../shared/components/Markdown';
 import { IdLookup } from '../shared/types/docs';
