+++ src/styles/components/_components.scss
@@ -12,6 +12,7 @@
 @import 'modal';
 @import 'nav-primary/nav-primary';
 @import 'nav-secondary';
+@import 'nav-tertiary';
 @import 'search';
 @import 'form';
 @import 'image';
